+++ frontend/src/context/AuthContext.js
@@ -1,4 +1,8 @@
+//************************** IMPORTED
+//REACT HOOKS/IMPORTS
 import { createContext, useReducer, useEffect } from "react";
+//**************************************************************
+
 export const AuthContext = createContext();
 
 export const authReducer = (state, action) => {
