+++ frontend/src/hooks/useEmployeeContext.js
@@ -1,5 +1,9 @@
+//************************** IMPORTED
+//REACT HOOKS/IMPORTS
 import { useContext } from "react";
+//CONTEXT
 import { EmployeeContext } from "../context/EmployeeContext";
+//**************************************************************
 
 export const useEmployeeContext = () => {
   const context = useContext(EmployeeContext);
