+++ frontend/src/hooks/useSelectContext.js
@@ -1,5 +1,9 @@
+//************************** IMPORTED
+//REACT HOOKS/IMPORTS
 import { useContext } from "react";
+//CONTEXT
 import { SelectContext } from "../context/SelectContext";
+//**************************************************************
 
 export const useSelectContext = () => {
   const context = useContext(SelectContext);
