+++ frontend/src/hooks/useSignup.js
@@ -1,8 +1,13 @@
+//************************** IMPORTED
+//REACT HOOKS/IMPORTS
 import { useState } from "react";
+//CUSTOM HOOKS
 import { useAuthContext } from "./useAuthContext";
 import { useEmployeeContext } from "../hooks/useEmployeeContext";
 import { useEmployeeCrudContext } from "./useEmployeeCrudContext";
+//ENV VARIBLES
 const apiURL = process.env.REACT_APP_DEVURL;
+//**************************************************************
 
 export const useSignup = () => {
   const [error, setError] = useState(null);
