+++ frontend/src/pages/HomeGeneral.js
@@ -1,4 +1,8 @@
+//************************** IMPORTED
+//STYLESHEET
 import "../stylesheets/HomeGeneral.css";
+//**************************************************************
+
 const HomeGeneral = () => {
   return <div className="homeGeneral-mainContainer"></div>;
 };
